refactor: tests package imports for test modules and rm relative paths
See pytest docs for details:
https://docs.pytest.org/en/7.1.x/explanation/pythonpath.html
