+++ tests/categorical/multicategorical_test_data.py
