+++ tests/stats/test_acovf.py
