+++ tests/stats/test_diebold_mariano.py
