+++ tests/test_weights.py
